Update script.js
change to parseFloat

+++ script.js
@@ -7,8 +7,8 @@ var js_soma;
 function somarnumeros() {
 /* Entrada de Dados */
 
-js_primeironumero = parseInt(document.getElementById('j_primeironumero').value);
-js_segundonumero = parseInt(document.getElementById('j_segundonumero').value);
+js_primeironumero = parseFloat(document.getElementById('j_primeironumero').value);
+js_segundonumero = parseFloat(document.getElementById('j_segundonumero').value);
 /* Processamento de Dados */
 js_soma= js_primeironumero + js_segundonumero;
 /* Saída de Dados */
